refactor dataset
