+++ tamr_unify_client/dataset/collection.py
@@ -1,9 +1,9 @@
 from tamr_unify_client.base_collection import BaseCollection
-from tamr_unify_client.models.dataset.resource import Dataset
+from tamr_unify_client.dataset.resource import Dataset
 
 
 class DatasetCollection(BaseCollection):
-    """Collection of :class:`~tamr_unify_client.models.dataset.resource.Dataset` s.
+    """Collection of :class:`~tamr_unify_client.dataset.resource.Dataset` s.
 
     :param client: Client for API call delegation.
     :type client: :class:`~tamr_unify_client.Client`
@@ -22,7 +22,7 @@ class DatasetCollection(BaseCollection):
         :param resource_id: The resource ID. E.g. ``"1"``
         :type resource_id: str
         :returns: The specified dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         return super().by_resource_id("datasets", resource_id)
 
@@ -32,7 +32,7 @@ class DatasetCollection(BaseCollection):
         :param relative_id: The resource ID. E.g. ``"datasets/1"``
         :type relative_id: str
         :returns: The specified dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         return super().by_relative_id(Dataset, relative_id)
 
@@ -42,7 +42,7 @@ class DatasetCollection(BaseCollection):
         :param external_id: The external ID.
         :type external_id: str
         :returns: The specified dataset, if found.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         :raises KeyError: If no dataset with the specified external_id is found
         :raises LookupError: If multiple datasets with the specified external_id are found
         """
@@ -53,7 +53,7 @@ class DatasetCollection(BaseCollection):
         over this collection.
 
         :returns: Stream of datasets.
-        :rtype: Python generator yielding :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: Python generator yielding :class:`~tamr_unify_client.dataset.resource.Dataset`
 
         Usage:
             >>> for dataset in collection.stream(): # explicit
@@ -69,7 +69,7 @@ class DatasetCollection(BaseCollection):
         :param dataset_name: Name of the desired dataset.
         :type dataset_name: str
         :return: Dataset with matching name in this collection.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         :raises KeyError: If no dataset with specified name was found.
         """
         for dataset in self:
@@ -84,7 +84,7 @@ class DatasetCollection(BaseCollection):
         :param creation_spec: Dataset creation specification should be formatted as specified in the `Public Docs for Creating a Dataset <https://docs.tamr.com/reference#create-a-dataset>`_.
         :type creation_spec: dict[str, str]
         :returns: The created Dataset
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         data = self.client.post(self.api_path, json=creation_spec).successful().json()
         return Dataset.from_json(self.client, data)
