+++ tamr_unify_client/models/project/mastering.py
@@ -1,5 +1,5 @@
+from tamr_unify_client.dataset.resource import Dataset
 from tamr_unify_client.models.binning_model import BinningModel
-from tamr_unify_client.models.dataset.resource import Dataset
 from tamr_unify_client.models.machine_learning_model import MachineLearningModel
 from tamr_unify_client.models.project.cluster_configuration import (
     PublishedClustersConfiguration,
@@ -15,11 +15,11 @@ class MasteringProject(Project):
         """Record pairs generated by Unify's binning model.
         Pairs are displayed on the "Pairs" page in the Unify UI.
 
-        Call :func:`~tamr_unify_client.models.dataset.resource.Dataset.refresh` from
+        Call :func:`~tamr_unify_client.dataset.resource.Dataset.refresh` from
         this dataset to regenerate pairs according to the latest binning model.
 
         :returns: The record pairs represented as a dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         alias = self.api_path + "/recordPairs"
         return Dataset(self.client, None, alias)
@@ -45,12 +45,12 @@ class MasteringProject(Project):
         High-impact pairs are displayed with a ⚡ lightning bolt icon on the
         "Pairs" page in the Unify UI.
 
-        Call :func:`~tamr_unify_client.models.dataset.resource.Dataset.refresh` from
+        Call :func:`~tamr_unify_client.dataset.resource.Dataset.refresh` from
         this dataset to produce new high-impact pairs according to the latest
         pair-matching model.
 
         :returns: The high-impact pairs represented as a dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         alias = self.api_path + "/highImpactPairs"
         return Dataset(self.client, None, alias)
@@ -60,11 +60,11 @@ class MasteringProject(Project):
         model. These clusters populate the cluster review page and get transient
         cluster ids, rather than published cluster ids (i.e., "Permanent Ids")
 
-        Call :func:`~tamr_unify_client.models.dataset.resource.Dataset.refresh` from
+        Call :func:`~tamr_unify_client.dataset.resource.Dataset.refresh` from
         this dataset to generate clusters based on to the latest pair-matching model.
 
         :returns: The record clusters represented as a dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         alias = self.api_path + "/recordClusters"
         return Dataset(self.client, None, alias)
@@ -73,7 +73,7 @@ class MasteringProject(Project):
         """Published record clusters generated by Unify's pair-matching model.
 
         :returns: The published clusters represented as a dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
 
         unified_dataset = self.unified_dataset()
@@ -104,7 +104,7 @@ class MasteringProject(Project):
         """Retrieves published cluster IDs for this project.
 
         :returns: The published cluster ID dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         # Replace this workaround with a direct API call once API
         # is fixed. APIs that need to work are: fetching the dataset and
@@ -121,7 +121,7 @@ class MasteringProject(Project):
         """Retrieves published cluster stats for this project.
 
         :returns: The published cluster stats dataset.
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         # Replace this workaround with a direct API call once API
         # is fixed. APIs that need to work are: fetching the dataset and
@@ -149,7 +149,7 @@ class MasteringProject(Project):
         """Project's unified dataset with associated clusters.
 
         :returns: The record clusters with data represented as a dataset
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype: :class:`~tamr_unify_client.dataset.resource.Dataset`
         """
         unified_dataset = self.unified_dataset()
 
@@ -165,7 +165,7 @@ class MasteringProject(Project):
     def published_clusters_with_data(self):
         """Project's unified dataset with associated clusters.
         :returns: The published clusters with data represented as a dataset
-        :rtype :class `~tamr_unify_client.models.dataset.resource.Dataset`
+        :rtype :class `~tamr_unify_client.dataset.resource.Dataset`
         """
 
         unified_dataset = self.unified_dataset()
