+++ tests/unit/test_dataset_geo.py
@@ -8,7 +8,7 @@ import responses
 
 from tamr_unify_client import Client
 from tamr_unify_client.auth import UsernamePasswordAuth
-from tamr_unify_client.models.dataset.resource import Dataset
+from tamr_unify_client.dataset.resource import Dataset
 
 
 class TestDatasetGeo(TestCase):
